Schema and DataStructure functionality is unified
