+++ lib/schema/controls.rb
@@ -4,6 +4,5 @@ require 'schema/controls/random'
 require 'schema/controls/attribute'
 require 'schema/controls/schema'
 require 'schema/controls/schema/different'
-require 'schema/controls/data_structure'
 require 'schema/controls/comparison'
 require 'schema/controls/comparison/entry'
