+++ test/automated/schema/data_structure.rb
@@ -2,12 +2,12 @@ require_relative '../automated_init'
 
 context "Data Structure" do
   test "Is a Schema object" do
-    ancestors = Schema::Controls::DataStructure.ancestors
+    ancestors = Schema::Controls::Schema.ancestors
     assert(ancestors.include? Schema)
   end
 
   test "Can configure its dependencies" do
-    example = Schema::Controls::DataStructure::ConfigureDependencies.example
+    example = Schema::Controls::Schema::ConfigureDependencies.example
     assert(example.some_dependency == :set)
   end
 end
