+++ test/automated/schema/strictness.rb
@@ -4,8 +4,8 @@ context "Data Structure" do
   context "Strictness" do
     context "Hash with extra attributes" do
       context "Not Strict (Default)" do
-        data = Schema::Controls::DataStructure::ExtraAttributes.data
-        data_structure = Schema::Controls::DataStructure::Example.build(data)
+        data = Schema::Controls::Schema::ExtraAttributes.data
+        data_structure = Schema::Controls::Schema::ExtraAttributes::Example.build(data)
 
         test "Sets matching attributes" do
           assert(data_structure.some_attribute == 'some value')
@@ -15,11 +15,11 @@ context "Data Structure" do
       context "Strict" do
         strict = true
 
-        data = Schema::Controls::DataStructure::ExtraAttributes.data
+        data = Schema::Controls::Schema::ExtraAttributes.data
 
         test "Is incorrect" do
           assert_raises(Schema::Error) do
-            Schema::Controls::DataStructure::Example.build(data, strict)
+            Schema::Controls::Schema::ExtraAttributes::Example.build(data, strict)
           end
         end
       end
