+++ test/automated/schema/transform/transform_out_alias.rb
@@ -2,7 +2,7 @@ require_relative '../../automated_init'
 
 context "Schema" do
   context "Transform Out Alias" do
-    data_structure = Schema::Controls::DataStructure::ReadAndWrite.example
+    data_structure = Schema::Controls::Schema::ReadAndWrite.example
 
     refute(data_structure.some_attribute == 'some written value')
